HW13 refactoring
